Began more flexible multi entry XML data file loading, still needs class templates

+++ Khazad/src/Core/main.cpp
@@ -227,7 +227,7 @@ int main(int argv, char** argc)
 
 		FPSTimer->Pause(); // FrameRate Captures whole loop
 
-        FrameRate = FrameRateControl(FPSTimer);
+        FrameRate = (Uint32) FrameRateControl(FPSTimer);
 
 	} // Program done, exited
